Move connection.js to http directory.
