+++ src/core/store/useEntities.js
@@ -12,7 +12,7 @@ import { getRecords } from '../idb';
 import { cacheEntity } from '../idb/cache';
 import flattenEntity from '../utils/flattenEntity';
 import asArray from '../utils/asArray';
-import { STATUS_IN_PROGRESS, updateStatus } from './connection';
+import { STATUS_IN_PROGRESS, updateStatus } from '../http/connection';
 import { alert } from './alert';
 import interceptor from '../http/interceptor';
 import { PromiseQueue } from '../utils/promises';
